TZE validation test for demo extension now passes.

+++ src/rust/include/librustzcash.h
@@ -354,12 +354,12 @@ extern "C" {
 
     bool librustzcash_tze_verify(
         uint32_t cbranch,
-        uint64_t p_extension_id, // TODO: maybe this should use uint32_t all the way through
-        uint64_t p_mode, 
+        uint32_t p_extension_id, // TODO: maybe this should use uint32_t all the way through
+        uint32_t p_mode, 
         const unsigned char* p_payload,
         size_t p_size,
-        uint64_t w_extension_id,
-        uint64_t w_mode,
+        uint32_t w_extension_id,
+        uint32_t w_mode,
         const unsigned char* w_payload,
         size_t w_size,
         int32_t height,
